fix(paywall): keep the free plan door out of a subscriber's reach
Review follow-ups:

- The endpoint only answers someone the paywall is actually holding. It
  had no plan check, so a Pro subscriber with a stale paywall tab could
  disconnect their own banks and keep being billed for the sync.
  `subscriptions.enabled` gates it too, like everything else here.
- A former subscriber past the delay is offered one way out, not two: the
  manual trip to Settings → Connections stays only while the confirmed
  route is still held back.
- The modal describes the free plan rather than taking inventory of the
  account: the hard gate is reached with a bank *or* AI on, so "we switch
  all three off" claimed features the user may never have turned on.

+++ app/Models/User.php
@@ -143,7 +143,7 @@ class User extends Authenticatable implements HasLocalePreference, MustVerifyEma
         }
 
         return $this->onboarded_at
-            ->addHours((int) config('subscriptions.free_plan_escape_delay_hours'))
+            ->addHours(config('subscriptions.free_plan_escape_delay_hours'))
             ->isPast();
     }
